test: add integration tests for MatchesController. Accomplishments: - Implemented GET /api/matches/{id} integration test. - Implemented PUT /api/matches/{id}/result integration tests for success and authorization scenarios. - Added SeedTeamRegistrationAsync helper to handle playerrosters database constraints. - Resolved C# 12 primary constructor capture warning by explicitly capturing ITestOutputHelper.

+++ TTA.DataAccess/SQLScripts/03-Storage-procedures.sql
@@ -682,7 +682,7 @@ CREATE OR REPLACE FUNCTION public.upsert_match(
     p_scheduled_at TIMESTAMPTZ,
     p_match_number VARCHAR,
     p_venue VARCHAR,
-    p_temperature DOUBLE PRECISION,
+    p_temperature FLOAT,
     p_home_score INT,
     p_guest_score INT,
     p_created_at TIMESTAMPTZ
